# services/abysscore/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API of AbyssCore server",
        "title": "AbyssCore API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/items": {
            "get": {
                "description": "Returns a paginated list of game items with sorting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Items"
                ],
                "summary": "List game items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "name",
                            "collection",
                            "type",
                            "rarity"
                        ],
                        "type": "string",
                        "description": "Field to sort by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "order_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of game items",
                        "schema": {
                            "$ref": "#/definitions/examples.PaginatedGameItemsDTOResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid query params",
                        "schema": {
                            "$ref": "#/definitions/examples.BadRequestResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Admin creates a new game item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Items"
                ],
                "summary": "Create game item",
                "parameters": [
                    {
                        "description": "Game item data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateUpdateGameItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created game item",
                        "schema": {
                            "$ref": "#/definitions/examples.CreateGameItemDTOSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - missed request fields",
                        "schema": {
                            "$ref": "#/definitions/examples.BadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not enough rights",
                        "schema": {
                            "$ref": "#/definitions/examples.ForbiddenByAccessLevelResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - invalid request types",
                        "schema": {
                            "$ref": "#/definitions/examples.UnprocessableEntityResponse"
                        }
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "description": "Returns game item by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Items"
                ],
                "summary": "Get game item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game item",
                        "schema": {
                            "$ref": "#/definitions/examples.FindGameItemDTOSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "$ref": "#/definitions/examples.BadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not found - no such game item",
                        "schema": {
                            "$ref": "#/definitions/examples.GameItemNotFound"
                        }
                    }
                }
            },
            "put": {
                "description": "Admin updates a game item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Items"
                ],
                "summary": "Update game item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated game item data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateUpdateGameItem"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Game item updated"
                    },
                    "400": {
                        "description": "Bad request - missed request fields",
                        "schema": {
                            "$ref": "#/definitions/examples.BadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not enough rights",
                        "schema": {
                            "$ref": "#/definitions/examples.ForbiddenByAccessLevelResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - invalid request types",
                        "schema": {
                            "$ref": "#/definitions/examples.UnprocessableEntityResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Admin deletes a game item by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game Items"
                ],
                "summary": "Delete game item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Game item deleted"
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "$ref": "#/definitions/examples.BadRequestResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not enough rights",
                        "schema": {
                            "$ref": "#/definitions/examples.ForbiddenByAccessLevelResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - invalid request types",
                        "schema": {
                            "$ref": "#/definitions/examples.UnprocessableEntityResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.GameItemDTO": {
            "type": "object",
            "properties": {
                "collection": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rarity": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "examples.BadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "bad request"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "examples.CreateGameItemDTOSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/dto.GameItemDTO"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "examples.FindGameItemDTOSuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/dto.GameItemDTO"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "examples.ForbiddenByAccessLevelResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 403
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "insufficient access level"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "examples.GameItemNotFound": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "game item not found"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "examples.PaginatedGameItemsDTOResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GameItemDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "examples.UnprocessableEntityResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 422
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "unprocessable entity"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "request.CreateUpdateGameItem": {
            "type": "object",
            "required": [
                "collection",
                "name",
                "rarity",
                "type"
            ],
            "properties": {
                "collection": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rarity": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
    }
}